Remove unused SmartContract classes
